+++ plugins/e_menu/e_menu/src/apaveinput.tcl
@@ -34,7 +34,7 @@
 
 package require Tk
 
-package provide apave 2.9b3
+package provide apave 2.9b4
 
 source [file join [file dirname [info script]] apavedialog.tcl]