+++ plugins/e_menu/e_menu/src/pkgIndex.tcl
@@ -1,2 +1,2 @@
 
-package ifneeded apave 2.9b3 [list source [file join $dir apaveinput.tcl]]
+package ifneeded apave 2.9b4 [list source [file join $dir apaveinput.tcl]]